requirements.txt
mangum>=0.17.0
fastapi
uvicorn
python-dotenv
pydantic
httpx
stripe>=5.0.0
sqlalchemy>=2.0.0
pyjwt>=2.8.0